requirements.txt
flask
flask-cors
groq
gTTS
pdfplumber
python-dotenv
gunicorn